Calc transition STAKING_RATIO at runtime, make conversion ratio 64bits
For mocknet and testing we now allow changing the staking requirement
at runtime (i.e. for testing different staking requirements and
conversion ratios) the staking ratio needs to be updated with the
appropriate values.

Allow a higher precision conversion ratio to allow for more fine grained
conversions. For the $OXEN to $SESH conversion we now defer to a 128 bit
multiply to avoid overflow for the conversion.

+++ src/cryptonote_core/sent_transition/generate.py
@@ -166,7 +166,7 @@ print(
     f"""
 }};  // {len(oxen_sent)} registered addresses
 
-const std::pair<std::uint8_t, std::uint8_t> conv_ratio{{{args.ratio[0]}, {args.ratio[1]}}};
+const std::pair<std::uint64_t, std::uint64_t> conv_ratio{{{args.ratio[0]}, {args.ratio[1]}}};
 
 const std::unordered_map<std::string, std::uint64_t> transition_bonus{{
 """,
